fixes issues cause by previous attempt to resolve conflicts
