+++ tests/conftest.py
@@ -5,6 +5,7 @@ from .fixtures.fixtures_DataFrameWrapper import (  # noqa: F401
     expected_attrs,
     expected_dataframe,
     expected_dataframe_filter,
+    expected_dataframe_grouped,
     expected_dataframe_joined,
     expected_dataframe_limit,
     expected_dataframe_schema,
@@ -13,6 +14,7 @@ from .fixtures.fixtures_DataFrameWrapper import (  # noqa: F401
     expected_empty_metadata,
     expected_hl_dataset_no_metadata,
     expected_hl_dataset_with_metadata,
+    expected_schema_grouped,
     expected_schema_joined,
     expected_schema_select,
     hl_dataframe,
