+++ tests/fixtures/fixtures_DataFrameWrapper.py
@@ -10,6 +10,7 @@ from pyspark.sql import SparkSession
 from pyspark.sql.types import (
     ArrayType,
     BinaryType,
+    DoubleType,
     IntegerType,
     StringType,
     StructField,
@@ -43,6 +44,7 @@ def expected_dataframe_schema() -> StructType:
     """Schema for expected DataFrame."""
     return StructType(
         [
+            StructField("category", StringType(), True),
             StructField("a", IntegerType(), True),
             StructField("b", IntegerType(), True),
             StructField("c", IntegerType(), True),
@@ -60,9 +62,9 @@ def expected_dataframe(
     # `Value of type variable "RowLike" of "createDataFrame" of "SparkSession" cannot
     # be "Dict[str, int]"`
     data: Iterable = [
-        {"a": 1, "b": 4, "c": 7},
-        {"a": 2, "b": 5, "c": 8},
-        {"a": 3, "b": 6, "c": 9},
+        {"category": "x", "a": 1, "b": 4, "c": 7},
+        {"category": "x", "a": 2, "b": 5, "c": 8},
+        {"category": "y", "a": 3, "b": 6, "c": 9},
     ]
     return spark_session.createDataFrame(
         data=data,
@@ -97,8 +99,8 @@ def expected_dataframe_limit(
 ) -> SparkDataFrame:
     """Expected data when using limit call method."""
     data: Iterable = [
-        {"a": 1, "b": 4, "c": 7},
-        {"a": 2, "b": 5, "c": 8},
+        {"category": "x", "a": 1, "b": 4, "c": 7},
+        {"category": "x", "a": 2, "b": 5, "c": 8},
     ]
     return spark_session.createDataFrame(
         data=data,
@@ -141,7 +143,7 @@ def expected_dataframe_filter(
 ) -> SparkDataFrame:
     """Expected data when using filter call method."""
     data: Iterable = [
-        {"a": 3, "b": 6, "c": 9},
+        {"category": "y", "a": 3, "b": 6, "c": 9},
     ]
     return spark_session.createDataFrame(
         data=data,
@@ -184,6 +186,7 @@ def expected_schema_joined() -> StructType:
     return StructType(
         [
             StructField("a", IntegerType(), True),
+            StructField("category", StringType(), True),
             StructField("b", IntegerType(), True),
             StructField("c", IntegerType(), True),
             StructField("d", IntegerType(), True),
@@ -199,13 +202,45 @@ def expected_dataframe_joined(
 ) -> SparkDataFrame:
     """Expected DataFrame once `received` and `other` have been joined."""
     data: Iterable = [
-        {"a": 1, "b": 4, "c": 7, "d": 10, "e": 13},
-        {"a": 2, "b": 5, "c": 8, "d": 11, "e": 14},
-        {"a": 3, "b": 6, "c": 9, "d": 12, "e": 15},
+        {"a": 1, "category": "x", "b": 4, "c": 7, "d": 10, "e": 13},
+        {"a": 2, "category": "x", "b": 5, "c": 8, "d": 11, "e": 14},
+        {"a": 3, "category": "y", "b": 6, "c": 9, "d": 12, "e": 15},
     ]
     return spark_session.createDataFrame(
         data=data,
         schema=expected_schema_joined,
+    )
+
+
+@fixture
+def expected_schema_grouped() -> StructType:
+    """Schema for expected DataFrame."""
+    return StructType(
+        [
+            StructField("category", StringType(), True),
+            StructField("avg(a)", DoubleType(), True),
+            StructField("avg(b)", DoubleType(), True),
+            StructField("avg(c)", DoubleType(), True),
+        ]
+    )
+
+
+@fixture
+def expected_dataframe_grouped(
+    spark_session: SparkSession,
+    expected_schema_grouped: StructType,
+) -> SparkDataFrame:
+    """A dummy `DataFrame` for testing."""
+    # Annotating `data` with `Iterable` stop `mypy` from complaining that
+    # `Value of type variable "RowLike" of "createDataFrame" of "SparkSession" cannot
+    # be "Dict[str, int]"`
+    data: Iterable = [
+        {"category": "x", "avg(a)": 1.5, "avg(b)": 4.5, "avg(c)": 7.5},
+        {"category": "y", "avg(a)": 3.0, "avg(b)": 6.0, "avg(c)": 9.0},
+    ]
+    return spark_session.createDataFrame(
+        data=data,
+        schema=expected_schema_grouped,
     )
